Fix test_bounce_animation
When I ran my initial tox the `test_bounce_animation` test was
failing for me because it seemed to assume that my console would
be 80 characters wide, which it was not. I used your existing
`get_terminal_columns` function to remove the assumption, and now
the test should pass regardless of the width of the tester's
console.

I also fixed a typo.