Add tests for checking color of spinner in stdout
`_get_test_output` has been refactored to take a `no_ansi` arg,
which defaults to `True`. We then use this in the new test to keep
the ANSI escape characters and compare them to those supplied by
colorama.

It feels a little like we're testing another library with this
test but I can understand why you feel like you need to have it
in there.